+++ src/com/grivera/solver/Model.java
@@ -3,7 +3,7 @@ package com.grivera.solver;
 public interface Model {
     void run();
     void run(int episodes);
-    int getTotalCost();
-    int getTotalProfit();
+    long getTotalCost();
+    long getTotalProfit();
     void printRoute();
 }
